refactor: service, repo 합침
