+++ src/test/java/com/iabacus/salespro/web/project/service/ContractServiceTest.java
@@ -12,7 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import com.iabacus.salespro.web.IntegrationTestSupport;
 import com.iabacus.salespro.web.aggregate.domain.MonthlyEmployeeCostAggregate;
-import com.iabacus.salespro.web.aggregate.repository.MonthlyEmployeeCostAggregateRepository;
+import com.iabacus.salespro.web.aggregate.repository.AggregateRepository;
 import com.iabacus.salespro.web.common.Money;
 import com.iabacus.salespro.web.common.Ratio;
 import com.iabacus.salespro.web.employee.domain.Employee;
@@ -45,7 +45,7 @@ class ContractServiceTest extends IntegrationTestSupport {
     private ProjectRepository projectRepository;
 
     @Autowired
-    private MonthlyEmployeeCostAggregateRepository monthlyEmployeeCostAggregateRepository;
+    private AggregateRepository aggregateRepository;
 
     @Autowired
     private ContractService contractService;
@@ -117,7 +117,7 @@ class ContractServiceTest extends IntegrationTestSupport {
         List<Input> inputs = inputRepository.findByContractId(contract.getId());
         assertThat(inputs.size()).isEqualTo(2);
 
-        List<MonthlyEmployeeCostAggregate> monthlyEmployeeCostAggregates = monthlyEmployeeCostAggregateRepository.findAll();
+        List<MonthlyEmployeeCostAggregate> monthlyEmployeeCostAggregates = aggregateRepository.findAll();
         assertThat(monthlyEmployeeCostAggregates.size()).isEqualTo(4);
     }
