+++ src/test/java/com/iabacus/salespro/web/project/service/InputServiceTest.java
@@ -12,8 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import com.iabacus.salespro.web.IntegrationTestSupport;
 import com.iabacus.salespro.web.aggregate.domain.MonthlyEmployeeCostAggregate;
-import com.iabacus.salespro.web.aggregate.repository.MonthlyEmployeeCostAggregateRepository;
-import com.iabacus.salespro.web.aggregate.service.SalesAggregateService;
+import com.iabacus.salespro.web.aggregate.repository.AggregateRepository;
 import com.iabacus.salespro.web.common.Money;
 import com.iabacus.salespro.web.common.Ratio;
 import com.iabacus.salespro.web.employee.domain.Employee;
@@ -48,7 +47,7 @@ class InputServiceTest extends IntegrationTestSupport {
     private InputRepository inputRepository;
 
     @Autowired
-    private MonthlyEmployeeCostAggregateRepository monthlyEmployeeCostAggregateRepository;
+    private AggregateRepository aggregateRepository;
 
     @Autowired
     private SalesAggregateService salesAggregateService;
@@ -118,7 +117,7 @@ class InputServiceTest extends IntegrationTestSupport {
 
         // 집계 데이터 4개 생성
         // 이동욱 - 1월, 2월 / 이지수 - 1월, 2월
-        List<MonthlyEmployeeCostAggregate> monthlyEmployeeCostAggregateList = monthlyEmployeeCostAggregateRepository.findAll();
+        List<MonthlyEmployeeCostAggregate> monthlyEmployeeCostAggregateList = aggregateRepository.findAll();
         assertThat(monthlyEmployeeCostAggregateList.size()).isEqualTo(4); // todo: 월단위로 쪼개서 4로 변해야함
     }
